Room 1-Absorb(Rough)
I put the enemy prefabs in Room 1 Absorb. The idea is that in this room, the player gets to practice the absorb mechanic. The first two enemies will be still. The next can also be still or possibly larger to make a slight increase in difficulty, then the last enemy is the big one which they can absorb. We can also change the room order too if this doesn't make sense mechanic and health bar wise. The black triangles are "trees".