added prefabs but also mistake
Added absorb bar and tried to add another fill in the background of the absorb bar so when absorb meter goes down you see black in the absorb bar, it is not showing up though. Also, originally I tried putting in an image for this second fill, but then that didn't work. Now there is a white image box that i can't find in the hierarchy and can't delete, so hopegully that just wont be put into the scene,. i also put aliana's background into the scene to show the color's im using.