created assets for control prompts and start prefab
Put in assets for some control prompts in their prefabs. I tried making the start button prefab have the image of the start button in it. It is there, but i am having trouble hiding the start game object color without moving it to the side of the image and making it go reverse.