Animations,AbsorbBar,Animationscript
These changes include animations of ekey prompt in the prefab. I created Keyprompt animations so that I could reuse it, and also created a script to animate it. I adjusted the AbsorbBar and made sure the slider is not checked as interactable. I put the new absorb bar in the test scene. The health bar in the test scene is trash, Ryan did some fixes to that in a different branch.