Updated UI prompts in lvl tst scene
I added the X key prompt into this, and this is where it should be for the other levels. For the prompts that are above the character, it really should just be one prompt at a time, but this is the general gist of placement. I hope this make sense! Also, I did put in the UI prefabs to the enemies as well, but I had a hard time seeing it on my screen and editing the sizes