Regular level UI done
regular level UI done for all of the control and key icons. In the places they are supposed to be. Also, i put in a background for the UI excluding the WASD keys. The animations for the UI may not be great, and I can fix that later for now but hopefully this helps for main  build!